create pfxPitcherMetrics store for pitch type percentile data Fixes #48
